exch.hpp adding switch for Msfield in E(), fixing warning for float to double cast

+++ src/llg_terms/micro_exch_nonequi.cpp
@@ -152,7 +152,7 @@ af::array NonequiExchangeField::calc_COO_matrix(const double A_exchange, const N
     double time_convert = timer_convert.stop();
 
     if(verbose) {
-        printf("%s Initialized sparse COO exchange matrix in %f [s]. Converted COO to CSR in %f [s]. Sparsity = %f\n", Info(), time, time_convert, (float)af::sparseGetNNZ(matr_CSR) / (float)matr_CSR.elements());
+        printf("%s Initialized sparse COO exchange matrix in %f [s]. Converted COO to CSR in %f [s]. Sparsity = %f\n", Info(), time, time_convert, static_cast<double>(af::sparseGetNNZ(matr_CSR)) / static_cast<double>(matr_CSR.elements()));
         fflush(stdout);
     }
     return matr_CSR;
@@ -314,7 +314,7 @@ af::array NonequiExchangeField::calc_COO_matrix(const af::array& A_exchange_fiel
     double time_convert = timer_convert.stop();
 
     if(verbose) {
-        printf("%s Initialized sparse COO exchange matrix in %f [s]. Converted COO to CSR in %f [s]. Sparsity = %f\n", Info(), time, time_convert, (float)af::sparseGetNNZ(matr_CSR) / (float)matr_CSR.elements());
+        printf("%s Initialized sparse COO exchange matrix in %f [s]. Converted COO to CSR in %f [s]. Sparsity = %f\n", Info(), time, time_convert, static_cast<double>(af::sparseGetNNZ(matr_CSR)) / static_cast<double>(matr_CSR.elements()));
         fflush(stdout);
     }
 
@@ -421,7 +421,7 @@ af::array NonequiExchangeField::calc_CSR_matrix(const double A_exchange, const N
     }
 
     af::array result = af::sparse((dim_t) dimension, (dim_t) dimension, (dim_t) CSR_values.size(), (void*) CSR_values.data(), CSR_IA.data(), CSR_JA.data(), f64);
-    if(verbose) printf("%s Initialized sparse exchange matrix in %f [s]. Sparsity of CSR_matrix = %f\n", Info(), t.stop(), (float)af::sparseGetNNZ(matr) / (float)matr.elements());
+    if(verbose) printf("%s Initialized sparse exchange matrix in %f [s]. Sparsity of CSR_matrix = %f\n", Info(), t.stop(), static_cast<double>(af::sparseGetNNZ(matr)) / static_cast<double>(matr.elements()));
     return result;
 }
 
@@ -583,7 +583,7 @@ af::array NonequiExchangeField::calc_CSR_matrix(const af::array& A_exchange_fiel
     af::freeHost(a_host);
     af::array result = af::sparse((dim_t) dimension, (dim_t) dimension, (dim_t) CSR_values.size(), (void*) CSR_values.data(), CSR_IA.data(), CSR_JA.data(), f64);
     if(verbose) {
-        printf("%s Initialized sparse exchange matrix in %f [s]. Sparsity of CSR_matrix = %f\n", Info(), t.stop(), (float)af::sparseGetNNZ(result) / (float)result.elements());
+        printf("%s Initialized sparse exchange matrix in %f [s]. Sparsity of CSR_matrix = %f\n", Info(), t.stop(), static_cast<double>(af::sparseGetNNZ(result)) / static_cast<double>(result.elements()));
         fflush(stdout);
     }
     return result;
