+++ src/llg_terms/micro_exch_rkky.cpp
@@ -172,7 +172,7 @@ af::array RKKYExchangeField::calc_CSR_matrix(const af::array& RKKY_field, const 
     af::freeHost(rkky_indices_raw);
     af::array result = af::sparse((dim_t) dimension, (dim_t) dimension, (dim_t) CSR_values.size(), (void*) CSR_values.data(), CSR_IA.data(), CSR_JA.data(), f64);
     if(verbose) {
-        printf("%s Initialized sparse CSR RKKY-exchange matrix in %f [s]. Sparsity of CSR_matrix = %f\n", Info(), t.stop(), (float)af::sparseGetNNZ(result) / (float)result.elements());
+        printf("%s Initialized sparse CSR RKKY-exchange matrix in %f [s]. Sparsity of CSR_matrix = %f\n", Info(), t.stop(), static_cast<double>(af::sparseGetNNZ(result)) / static_cast<double>(result.elements()));
         fflush(stdout);
     }
     return result;
@@ -308,7 +308,7 @@ af::array RKKYExchangeField::calc_COO_matrix(const af::array& RKKY_field, const 
     double time_convert = timer_convert.stop();
 
     if(verbose) {
-        printf("%s Finished sparse matrix setup in %f [s]. Initialized COO matrix in %f [s]. Converted COO to CSR format in %f [s]. Sparsity = %f\n", Info(), time + time_convert, time, time_convert, (float)af::sparseGetNNZ(matr_CSR) / (float)matr_CSR.elements());
+        printf("%s Finished sparse matrix setup in %f [s]. Initialized COO matrix in %f [s]. Converted COO to CSR format in %f [s]. Sparsity = %f\n", Info(), time + time_convert, time, time_convert, static_cast<double>(af::sparseGetNNZ(matr_CSR)) / static_cast<double>(matr_CSR.elements()));
         fflush(stdout);
     }
